rename "parse-capi" to "model"
