+++ packages/frontend/lib/model/extract-capi.ts
