+++ packages/frontend/lib/model/extract-config.ts
