+++ packages/frontend/lib/model/extract-ga.ts
